+++ R/globals.R
@@ -1,0 +1,10 @@
+# R/globals.R
+
+utils::globalVariables(c(
+  "Affectation", "Valeur_Cadastrale", "Valeur_totale", "Estimation_impot",
+  "Nature", "med", "y_pos", "Commune", "Impot_total", "Groupe",
+  "Total_impot", "Pourcentage", "Nb_parcelles", "Etiquette", "Type_Bien",
+  "Prefecture", "Nb", "Pourcent", "ymax", "ymin", "labelPosition", "label",
+  "Localite", "Estimation_Impot", "Total_Impot", "Total_Valeur",
+  "Nb_Immeubles", "Total_Impot_k", "Total_Valeur_M"
+))